Replacing the 500k programs by 1m, and better folder structure.
The old collection used an old version of our extraction-reconstruction
framework, which was less efficient and did not preserve the original
folder structure of the extracted repositories. We performed an experiment
to extract and reconstruct 1 million programs, while keeping the directory
tree of the original repositories intact. Since this is a larger and more
organized version of the same benchmarks, we're replacing the old version
with this new one.